Tanda 58: zx_port_queue materializaba tres enums de Rust del que llama
Continuacion de la lente de la tanda 57: una estructura que se lee de la
memoria del proceso con UserInPtr no puede tener tipos con valores
restringidos. sys_port_queue le pasa el buffer del que llama a
UserInPtr::<PortPacket>::read, y PortPacket tenia tres.

- type_ era PacketType, un #[repr(u32)] con nueve variantes y un hueco: no
  hay tipo 8. status era ZxError, un #[repr(i32)] disperso entre -76 y 0.
  Materializar cualquiera de los dos a partir de un u32 o un i32 elegido por
  el proceso es comportamiento indefinido.
- PayloadRepr::decode tenia `_ => unimplemented!()`, y PageRequest es una
  variante valida que un proceso puede escribir: encolar un paquete de tipo
  9 panicaba el kernel. El camino es el info! de sys_port_queue, que formatea
  el paquete antes de que nada lo valide.
- PacketGuestVcpu guarda su propio discriminante dentro de Payload, o sea
  que ese byte tambien es del que llama, y los dos sitios que lo leen
  (PartialEq y Debug) hacian match sobre el con dos brazos y ninguno mas.

PortPacket pasa a ser la estructura del cable, con type_: u32 y status: i32
y el mismo repr(C) y los mismos 48 bytes, asi que zircon-syscall no se toca.
La conversion la hacen PacketType::from_raw y PacketGuestVcpuType::from_raw,
escritas a mano para que el hueco del 8 sea explicito. PayloadRepr::decode
devuelve ZxResult y su match ya no tiene brazo comodin. Debug de PortPacket
no puede fallar. El status viaja verbatim, como en Zircon.

Y el mismo syscall por el otro topico: zx_job_set_policy con
ZX_JOB_POL_TIMER_SLACK lee un TimerSlackPolicy de la memoria del que llama,
cuyo default_mode era un Slack de tres variantes que nadie comprobaba;
check_timer_policy solo miraba que min_slack no fuera negativo. Zircon
comprueba los dos, y zx_timer_create, la otra puerta al mismo enum, siempre
lo ha hecho. Ahora TimerSlackPolicy::parse valida los dos y devuelve lo que
dice, y generate_new recibe valores ya comprobados.

12 tests nuevos. zircon-object pasa de 177 a 189.

+++ zircon-object/src/hypervisor/vcpu.rs
@@ -131,7 +131,7 @@ impl TryInto<PortPacket> for rvm::RvmExitPacket {
         };
         Ok(PortPacketRepr {
             key: self.key,
-            status: ZxError::OK,
+            status: ZxError::OK as i32,
             data,
         }
         .into())
